get location page halfway working and refactored
