+++ client/src/components/userpage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Row from './row.jsx';
+import PicRow from './picrow.jsx';
 import axios from 'axios';
 
 class Userpage extends React.Component {
@@ -12,7 +12,7 @@ class Userpage extends React.Component {
       <div style={{minHeight: `calc(100vh - 150px)`}}>
         <h1 style={{fontFamily: `billabong`, textAlign: `center`, color: `#32bfff`}}>Hello {this.props.userData.firstName}</h1>
         <h2 style={{fontFamily: `billabong`, textAlign: `center`, color: `#919295`}}>Your Places</h2>
-        <Row rowType={"user"} data={this.props.userData} />
+        <PicRow rowType={"user"} data={this.props.userData} />
         <br />
       </div>
     )
