+++ client/src/components/worthymap.jsx
@@ -1,134 +1,147 @@
 import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import _ from 'lodash';
-import ClusteredMap from './clusteredmap.jsx';
+import { compose, withProps, withHandlers, lifecycle } from 'recompose';
+import { withScriptjs, withGoogleMap, GoogleMap, Marker } from 'react-google-maps';
+import { Grid, Row } from 'react-bootstrap';
+import { MarkerClusterer }  from 'react-google-maps/lib/components/addons/MarkerClusterer';
+import { SearchBox } from "react-google-maps/lib/components/places/SearchBox";
 
-class WorthyMap extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      markers: [],
-      userMarker: [],
-      zoom: 4,
-      position: {lat: 37.09, lng: -95.71}
-    };
-    // zoom: props.zoom,
-    // center: props.position
 
-    this.getLocations = this.getLocations.bind(this);
-    this.clickMap = this.clickMap.bind(this);
-  }
-  componentWillMount() {
-    this.setState({ markers: [] })
-  }
-
-  componentDidMount() {
-    this.getUserLocation();
-    this.getLocations();
-  }
-
-  getUserLocation() {
-    const onSuccess = ({coords}) => {
-      console.log('users coords', coords);
-      
+// https://tomchentw.github.io/react-google-maps/#markerclusterer
+const WorthyMap = compose(
+  withProps({
+    googleMapURL: "https://maps.googleapis.com/maps/api/js?key=AIzaSyCWb-bdFbk9oBVzcehnMVyaG1QqVCvf8As&v=3.exp&libraries=geometry,drawing,places",
+    loadingElement: <div style={{ height: `100%` }} />,
+    containerElement: <div style={{ height: `100%`, width: `100%`, border: `1px solid grey`, borderRadius: `3px`}}/>,
+    mapElement: <div style={{ height: `100%` }} />,
+  }),
+  withHandlers({
+    onMarkerClustererClick: () => (markerClusterer) => {
+      const clickedMarkers = markerClusterer.getMarkers()
+    },
+  }),
+  lifecycle({
+    componentWillMount() {
+      const refs = {}
       this.setState({
-        position: {
-          lat: coords.latitude, 
-          lng: coords.longitude
+        zoom: null,
+        bounds: null,
+        center: null,
+        refCenter: {lat: () => 84, lng: () => -100},
+        onMapMounted: ref => {
+          refs.map = ref;
         },
-        zoom: 9
-      });
-    };
+        onBoundsChanged: () => {
+          this.setState({
+            bounds: refs.map.getBounds(),
+            refCenter: refs.map.getCenter(),
+          })
+        },
+        onSearchBoxMounted: ref => {
+          refs.searchBox = ref;
+        },
+        onPlacesChanged: () => {
+          const places = refs.searchBox.getPlaces();
+          const bounds = new google.maps.LatLngBounds();
 
-    // if we can't get users location we try again without high accuracy
-    const onError = (err) => {
-      navigator.geolocation.getCurrentPosition(
-        onSuccess, 
-        (err2) => console.log('error getting location', err2), 
-        {maximumAge: 3600000, timeout: 5000, enableHighAccuracy: false} 
-      );
-    };
+          places.forEach(place => {
+            if (place.geometry.viewport) {
+              bounds.union(place.geometry.viewport)
+            } else {
+              bounds.extend(place.geometry.location)
+            }
+          });
+          const nextMarkers = places.map(place => ({
+            position: place.geometry.location,
+          }));
+          const nextCenter = _.get(nextMarkers, '0.position', this.state.center);
 
-    // but first we try to get it with high accuracy
-    navigator.geolocation.getCurrentPosition(
-      onSuccess, 
-      onError, 
-      {maximumAge: 3600000, timeout: 5000, enableHighAccuracy: true} // maxAge is how old a cached result can be, timeout is how long to try getting position.
-    );
-  }
+          // hacky way of making sure the zoom updates even if the user changed the zoom
+          this.setState({
+            zoom: 1
+          })
+          this.setState({
+            center: nextCenter,
+            refCenter: nextCenter,
+            zoom: 11
+          });
+        },
+      })
+    },
+  }),
+  withScriptjs,
+  withGoogleMap
+) ((props) => {
+  const center = props.center || props.defaultCenter;
+  const zoom = props.zoom || props.defaultZoom;
+  const allMarkers = props.markers.map((marker, i) => (
+    <Marker
+      key={ i }
+      position={{ lat: marker.lat, lng: marker.lng }}
+      onClick={marker.clickHandler}
+    />
+  ));
+  const onCenterChanged = props.onCenterChanged ? props.onCenterChanged : () => {};
 
-  getLocations() {
-    if (!this.props.isForUploadPage) {
-      this.setState({markers: testingMarkers}) // delete this after getLocations is hooked up;
-      // TODO
-      // Make axios request to our server to get all the locations
-      // turn them all into makers and call setState and set an array of markers to this.state.markers
-    }
-  }
+  return (
+    <GoogleMap
+      ref={props.onMapMounted}
+      zoom={ zoom }
+      center={ center }
+      onBoundsChanged={ (...args) => {
+        props.onBoundsChanged(...args); 
+        onCenterChanged(props.refCenter.lat(), props.refCenter.lng());
+      } }
+      onClick={ props.clickMap }
+      onCenterChanged={ () => onCenterChanged(props.refCenter.lat(), props.refCenter.lng()) }
+    >
+      <MarkerClusterer
+        onClick={props.onMarkerClustererClick}
+        averageCenter
+        enableRetinaIcons
+        gridSize={60}
+      >
+        { allMarkers }
+      </MarkerClusterer>
+      <SearchBox
+        ref={props.onSearchBoxMounted}
+        bounds={props.bounds}
+        controlPosition={google.maps.ControlPosition.TOP_LEFT}
+        onPlacesChanged={props.onPlacesChanged}
+      >
+        <input
+          type="text"
+          placeholder="Search"
+          style={{
+            boxSizing: `border-box`,
+            border: `1px solid transparent`,
+            width: `240px`,
+            height: `32px`,
+            margin: 'auto auto auto auto',
+            marginTop: `10px`,
+            padding: `0 12px`,
+            borderRadius: `3px`,
+            boxShadow: `0 2px 6px rgba(0, 0, 0, 0.3)`,
+            fontSize: `14px`,
+            outline: `none`,
+            textOverflow: `ellipses`,
+            left: '500px'
+          }}
+        />
+      </SearchBox>
+    </GoogleMap>
+  );
+});
 
-  // allows a user to click the map to place a pin where there location is while uploading photos
-  clickMap(e) {
-    console.log(e);
-    this.props.getLocationUpload({
-      lat: e.latLng.lat(),
-      lng: e.latLng.lng()
-    });
-
-    this.setState({
-      userMarker: [{
-        lat: e.latLng.lat(),
-        lng: e.latLng.lng(),
-        clickHandler: () => {}
-      }]
-    })
-  }
-
-  render() {
-    const markers = this.state.markers.concat(this.state.userMarker);
-
-    let clickMap;
-    if (this.props.isForUploadPage) {
-      clickMap = this.clickMap;
-    } else {
-      clickMap = () => {};
-    }
-    console.log('renderprops', this.state.zoom, this.state.position)
-    return (
-      <ClusteredMap
-        markers={ markers } 
-        clickMap={ clickMap }
-        defaultZoom={ this.state.zoom }
-        defaultCenter={ this.state.position } 
-      />
-    );
-  }
+WorthyMap.propTypes = {
+  markers: PropTypes.array.isRequired,
+  clickMap: PropTypes.func,
+  defaultZoom: PropTypes.number.isRequired,
+  defaultCenter: PropTypes.shape({
+    lat: PropTypes.number,
+    lng: PropTypes.number
+  }).isRequired
 }
-
-
-const testingMarkers = [
-  {
-    lat: 34.05,
-    lng: -118.24,
-    clickHandler: function(e) {alert('i got clicked and my event is ' + JSON.stringify(e))}
-  },
-  {
-    lat: 34.25,
-    lng: -118.24,
-    clickHandler: function(e) {alert('i got clicked and my event is ' + JSON.stringify(e))}
-  },
-
-  {
-    lat: 34.35,
-    lng: -118.24,
-     clickHandler: function(e) {alert('i got clicked and my event is ' + JSON.stringify(e))}
-  },
-
-  {
-    lat: 34.45,
-    lng: -118.24,
-    clickHandler: function(e) {alert('i got clicked and my event is ' + JSON.stringify(e))}
-  }
-
-]
-
 
 export default WorthyMap;
