+++ package.json
@@ -46,10 +46,12 @@
     "passport": "^0.4.0",
     "passport-local": "^1.0.0",
     "passport-session": "^1.0.2",
+    "prop-types": "^15.6.1",
     "react": "^16.3.1",
     "react-bootstrap": "^0.32.1",
     "react-dom": "^16.3.1",
     "react-dropzone": "^4.2.9",
+    "react-fontawesome": "^1.6.1",
     "react-google-maps": "^9.4.5",
     "react-icons": "^2.2.7",
     "react-password-mask": "^3.3.0",
